+++ app/src/main/java/com/jianhui_zhu/simpleweatherwidget/dataprovider/model/No2.java
@@ -13,7 +13,7 @@ public class No2 implements Parcelable
 
     @SerializedName("v")
     @Expose
-    private int v;
+    private double v;
     public final static Parcelable.Creator<No2> CREATOR = new Creator<No2>() {
 
 
@@ -22,7 +22,7 @@ public class No2 implements Parcelable
         })
         public No2 createFromParcel(Parcel in) {
             No2 instance = new No2();
-            instance.v = ((int) in.readValue((int.class.getClassLoader())));
+            instance.v = ((double) in.readValue((double.class.getClassLoader())));
             return instance;
         }
 
@@ -38,7 +38,7 @@ public class No2 implements Parcelable
      * @return
      *     The v
      */
-    public int getV() {
+    public double getV() {
         return v;
     }
 
@@ -47,7 +47,7 @@ public class No2 implements Parcelable
      * @param v
      *     The v
      */
-    public void setV(int v) {
+    public void setV(double v) {
         this.v = v;
     }
