+++ app/src/main/java/com/jianhui_zhu/simpleweatherwidget/dataprovider/model/Wd.java
@@ -13,7 +13,7 @@ public class Wd implements Parcelable
 
     @SerializedName("v")
     @Expose
-    private int v;
+    private double v;
     public final static Parcelable.Creator<Wd> CREATOR = new Creator<Wd>() {
 
 
@@ -22,7 +22,7 @@ public class Wd implements Parcelable
         })
         public Wd createFromParcel(Parcel in) {
             Wd instance = new Wd();
-            instance.v = ((int) in.readValue((int.class.getClassLoader())));
+            instance.v = ((double) in.readValue((double.class.getClassLoader())));
             return instance;
         }
 
@@ -38,7 +38,7 @@ public class Wd implements Parcelable
      * @return
      *     The v
      */
-    public int getV() {
+    public double getV() {
         return v;
     }
 
@@ -47,7 +47,7 @@ public class Wd implements Parcelable
      * @param v
      *     The v
      */
-    public void setV(int v) {
+    public void setV(double v) {
         this.v = v;
     }
